Support isolated and fix wait
