+++ Doc/library/index.rst
@@ -77,4 +77,3 @@ the `Python Package Index <https://pypi.org>`_.
    unix.rst
    superseded.rst
    undoc.rst
-   _interpreters.rst
